create a pure pipeline
