+++ predict_prices/main.py
@@ -1,5 +1,4 @@
 from dataclasses import dataclass
-from typing import Iterable
 
 import lightgbm
 import numpy as np
@@ -7,57 +6,33 @@ import pandas as pd
 from sklearn import ensemble, linear_model  # type: ignore
 import xgboost
 
-import cross_validation
 import data_prep
+from predict_prices.hyperparameter_tuning import ModelHyperparameterMap
 
 
 @dataclass
-class ModelParameters:
-    model: any
-    categorical_encoders: data_prep.CategoricalEncoders
+class PurePipeline:
+    model_hyperparameter_map: ModelHyperparameterMap
+    train_data_prep_inputs: data_prep.TrainDataPrepInputs
+    test_df: pd.DataFrame
 
-@dataclass
-class ModelHyperparameterMap:
-    hyper_params_map: dict[any, dict[str, Iterable]]
+    def run(self):
+        # 1. perform data prep on training data
+        train_data_prep_outputs = self.train_data_prep_inputs.train_data_prep()
 
-    def measure_model_performance(
-            self,
-            train_data_prep_outputs: data_prep.TrainDataPrepOutputs,
-            n_iter: int,
-            folds: int
-    ) -> ModelParameters:
+        # 2. Measure performance of all of our different models
+        champ_parameters = self.model_hyperparameter_map.measure_model_performance(train_data_prep_outputs, n_iter, folds)
 
-        models = []
+        # 3. apply data prep from training to test data
+        test_data_prep_inputs = data_prep.TestDataPrepInputs(self.test_df, champ_parameters.categorical_encoders)
+        test_X = test_data_prep_inputs.test_data_prep()
 
-        for i in self.hyper_params_map:
-            param_grid = self.hyper_params_map[i]
+        # 4. predict test data
+        # the model was trained against log transformed target. Invert log for predictions
+        test_y = np.exp(champ_parameters.model.predict(test_X))
+        test_y = pd.DataFrame(test_y, columns=[self.train_data_prep_inputs.target_variable], index=test_X.index)
 
-            if len(param_grid) > 0:
-                best_model = cross_validation.bayes_cross_validation(i, train_data_prep_outputs.train_data, param_grid,
-                                                                     n_iter)
-                best_model_param_scores = cross_validation.save_model_performance_parameters(best_model, folds,
-                                                                                             train_data_prep_outputs)
-            else:
-                i.fit(train_data_prep_outputs.train_data.X, train_data_prep_outputs.train_data.y)
-                best_model_param_scores = cross_validation.save_model_performance_parameters(i, folds,
-                                                                                             train_data_prep_outputs)
-
-            models.append(best_model_param_scores)
-
-        param_scores = pd.DataFrame(models).sort_values('score', ascending=False)
-
-        print(param_scores[['model', 'score', 'standard_dev']])
-
-        # take only first row in case of ties (I don't care which model if they're tied)
-        champ = param_scores[param_scores.score == param_scores.score.max()].reset_index()
-
-        champ_model = champ.model[0]
-        champ_ord_enc = champ.ordinal_encoder[0]
-        champ_cat_enc = champ.categorical_encoder[0]
-
-        champ_parameters = ModelParameters(champ_model, data_prep.CategoricalEncoders(champ_ord_enc, champ_cat_enc))
-
-        return champ_parameters
+        return test_y
 
 
 def main():
@@ -110,27 +85,16 @@ def main():
     }
     hyper_param_map = ModelHyperparameterMap(model_dict)
 
-    #1. read in train and test data
+    # 1. read in train and test data
     train_df = data_prep.load_data(train_filepath)
     test_df = data_prep.load_data(test_filepath)
+    train_data_prep_inputs = data_prep.TrainDataPrepInputs(train_df, ohe_bool)
+    pure_pipeline_inputs = PurePipeline(hyper_param_map,train_data_prep_inputs, test_df)
 
-    # 2. perform data prep on training data
-    train_data_prep_inputs = data_prep.TrainDataPrepInputs(train_df, ohe_bool, target_variable)
-    train_data_prep_outputs = train_data_prep_inputs.train_data_prep()
+    # 2. run pure pipeline
+    test_y = pure_pipeline_inputs.run()
 
-    # 3. Measure performance of all of our different models
-    champ_parameters = hyper_param_map.measure_model_performance(train_data_prep_outputs, n_iter, folds)
-
-    # 4. apply data prep from training to test data
-    test_data_prep_inputs = data_prep.TestDataPrepInputs(test_df, champ_parameters.categorical_encoders)
-    test_X = test_data_prep_inputs.test_data_prep()
-
-    # 5. predict test data
-    # the model was trained against log transformed target. Invert log for predictions
-    test_y = np.exp(champ_parameters.model.predict(test_X))
-    test_y = pd.DataFrame(test_y, columns=[target_variable], index=test_X.index)
-
-    # 6. output predictions to csv
+    # 3. output predictions to csv
     test_y.to_csv('prediction.csv')
